Improvements to maintainability

+++ src/main/c++/irio_v2_cpp/utils.cpp
@@ -9,7 +9,7 @@ namespace utils{
 void throwIfNotSuccessNiFpga(const NiFpga_Status &status, const std::string &errMsg) {
 	if (NiFpga_IsError(status)) {
 		const std::string err = errMsg + std::string("(Code: ")
-				+ std::to_string(static_cast<std::int32_t>(status)) + std::string(")");
+				+ std::to_string(status) + std::string(")");
 		throw iriov2::errors::NiFpgaError(err);
 	}
 }
